:recycle: Move sopra_db.py to material_data folder
