+++ solcore/material_data/sopra_db.py
